requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.40
yfinance>=0.2.40
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
newspaper3k>=0.2.8
python-dotenv>=1.0.0
tavily-python>=0.3.0
duckduckgo-search>=4.1.0

